Add raftkv.
This builds but hasn't been tested. The main purpose of raftkv is to
allow writing a Jepsen test against our raft library, which is also
included. This has run (and passed) in the past, but the scripts need to
be updated with the new build system and paths before merging.