Switch curator database to flatbuffers
Some of the rationale for this change is in doc.go. In short,
flatbuffers allow fast read access to the database without copying,
allocating, or parsing unnecessary data. This is important in the
curator because there are several processes that repeatedly scan the
whole database, and this lets them do that much more efficiently.

Obviously, this is a backwards-incompatible change. If anyone has a
curator database they'd like to preserve, they can write a conversion
utility.

The changes in names of enum constants are disappointing, but I felt
like switching the enums to be generated by flatbuffers as well was
worth it for consistency.

Note that some of the flatbuffers generated files were generated with a
modified version of flatc and won't match what's generated by v1.10.0.
I've submitted PRs for the changes and I hope they'll be accepted.